Remove MenuLayout in favor of NativeWindowViews::Layout
